replace triggerVal with limiter where reasonable
